+++ IsekaiMod/Content/Features/IsekaiProtagonist/GodEmperor/Godhood.cs
@@ -4,6 +4,7 @@ using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
 using Kingmaker.Blueprints.Classes.Spells;
 using Kingmaker.Designers.Mechanics.Facts;
+using Kingmaker.Enums.Damage;
 using Kingmaker.UnitLogic.FactLogic;
 
 namespace IsekaiMod.Content.Features.IsekaiProtagonist.GodEmperor
@@ -15,12 +16,27 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.GodEmperor
             var Icon_Godhood = AssetLoader.LoadInternal("Features", "ICON_GODHOOD.png");
             var Godhood = Helpers.CreateBlueprint<BlueprintFeature>("Godhood", bp => {
                 bp.SetName("Godhood");
-                bp.SetDescription("At 20th level, you become a god. You gain 100 spell resistance and are immune to all {g|Encyclopedia:Physical_Damage}physical damage{/g}. "
-                    + "Your attacks ignore concealment and damage reduction. Any critical threats you make are automatically confirmed. "
+                bp.SetDescription("At 20th level, you become a god. You gain 100 spell resistance and are immune acid, cold, electricity, fire, sonic, and all "
+                    + "{g|Encyclopedia:Physical_Damage}physical damage{/g}. Your attacks ignore concealment and damage reduction. Any critical threats you make are automatically confirmed. "
                     + "The spells you cast ignore spell resistance and spell immunity.");
                 bp.m_Icon = Icon_Godhood;
                 bp.AddComponent<AddSpellResistance>(c => {
                     c.Value = 100;
+                });
+                bp.AddComponent<AddEnergyImmunity>(c => {
+                    c.Type = DamageEnergyType.Acid;
+                });
+                bp.AddComponent<AddEnergyImmunity>(c => {
+                    c.Type = DamageEnergyType.Cold;
+                });
+                bp.AddComponent<AddEnergyImmunity>(c => {
+                    c.Type = DamageEnergyType.Electricity;
+                });
+                bp.AddComponent<AddEnergyImmunity>(c => {
+                    c.Type = DamageEnergyType.Fire;
+                });
+                bp.AddComponent<AddEnergyImmunity>(c => {
+                    c.Type = DamageEnergyType.Sonic;
                 });
                 bp.AddComponent<AddPhysicalImmunity>();
                 bp.AddComponent<IgnoreConcealment>();