+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Effect Immunity` exceptional feat.
 ### Changed
 - Reworked `Isekai Protagonist` class progression to have more Overpowered Abilities instead of training arc feats.
+- Buffed `Godhood` to give energy immunity against acid, cold, electricity, fire, and sonic.
 - Nerfed `Grasp Heart` to stun if the creature fails a DC 50 fortitude save.
 - Changed `Glorious Aura` to give a +4 sacred bonus to all attributes instead.
 - Changed `Training Montage` to give a +10 untyped bonus to all attributes but now requires character level 10.